move code to model
